Add concourse-chrome-driver build to pipeline.
Add salesforce-splunk build to pipeline
Remove dependencies for splunk containers on base image build

I'm not sure about the &build-image-source anchor
I assume it's setting the creds everywhere but it's also
setting the repository which in most cases is wrong.
If we want to do that should we make a block that just does
the creds instead?