Get a list of changed files for the latest merged PR
Given a GitHub PAT org/repo and a target file name.

This will look for closed and merged PRs.
It will check that the merge target is main/master.
It will get the most recent PR matching those criteria.
It will then get a list of the files changed in that PR.
Then it outputs the list of changed file paths in JSON.